serialization::Array base class has more competence
