+++ src/serializers/json/jsondeserializerworker.cpp
@@ -98,35 +98,10 @@ TriggerPropertyDummyValueType JSONDeserializerWorker::get_trigger_dummy_value()
 
 std::unique_ptr<DeserializationArray> JSONDeserializerWorker::start_array()
 {
-  class JSONArray : public DeserializationArray
-  {
-  public:
-    explicit JSONArray(JSONDeserializerWorker& parent, const std::size_t size)
-        : DeserializationArray(parent)
-        , m_size(size)
-    {
-    }
-
-    [[nodiscard]] DeserializerWorker& next() override
-    {
-      m_current = m_parent.sub(m_next_index);
-      m_next_index += 1;
-      return *m_current;
-    }
-
-    [[nodiscard]] std::size_t size() const override
-    {
-      return m_size;
-    }
-
-  private:
-    const std::size_t m_size;
-  };
-
   if (!m_value.is_array()) {
     throw AbstractDeserializer::DeserializeError{"Expected Array"};
   }
-  return std::make_unique<JSONArray>(*this, m_value.size());
+  return std::make_unique<DeserializationArray>(*this, m_value.size());
 }
 
 std::unique_ptr<DeserializerWorker> JSONDeserializerWorker::sub(const std::string& key)
